refactor: rename utils/templates.ts to template.ts
